// apps/web/src/generated/types/experiments/GetExperiments.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


/**
 * @description Response for status 200
*/
export type GetExperiments200 = {
    /**
     * @type string
    */
    id: string;
    /**
     * @type string
    */
    title: string;
    /**
     * @type string,null
    */
    hypothesis: string | null;
    /**
     * @type string,null
    */
    status: string | null;
    /**
     * @type string
    */
    projectId: string;
    /**
     * @type string
    */
    projectName: string;
    /**
     * @type string,null
    */
    leadName: string | null;
    /**
     * @type number
    */
    measurementCount: number;
    /**
     * @type string,null
    */
    startDate: string | null;
}[];

export type GetExperimentsQueryResponse = GetExperiments200;

export type GetExperimentsQuery = {
    Response: GetExperiments200;
    Errors: any;
};