// apps/web/src/generated/types/samples/GetSamples.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


/**
 * @description Response for status 200
*/
export type GetSamples200 = {
    /**
     * @type string
    */
    id: string;
    /**
     * @type string
    */
    code: string;
    /**
     * @type string
    */
    specimenType: string;
    /**
     * @type string,null
    */
    collectedAt: string | null;
    /**
     * @type string,null
    */
    storageLocation: string | null;
    /**
     * @type number
    */
    experimentCount: number;
}[];

export type GetSamplesQueryResponse = GetSamples200;

export type GetSamplesQuery = {
    Response: GetSamples200;
    Errors: any;
};